requirements.txt
numpy==1.15.4
opencv_python==4.0.0.21
Flask_Cors==3.0.7
Keras==2.2.4
requests==2.21.0
imutils==0.5.2
Flask==1.0.2
Pillow==6.1.0
